Move most of RubricContainer to RubricContent
